Prettify config file
Put global and cow configuration into separate categories, so that it's easier to find the global config options. Otherwise, they're sorted into the middle of the cows.

Unfortunately, the breaks backwards compatibility with old config files. Forge does provide the means to track config version and rearrange properties, but that requires having included a version identifier in the past. I've set the config version to "1", to guard against the future.