Fix config comments not displaying
Reloading and saving the configuration in the middle of adding properties seems to break things, such as comments. It also runs the risk of losing portions of the config, if the game crashes in between updating sections.

It's not necessary to explicitly load the configuration file any more anyway -- forge handles it behind the scenes.